Porting: Data files renamed and C++ed
