+++ src/audio-sdlmixer.cc
@@ -30,9 +30,9 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/log.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/sfx2wav.h"
 #include "src/xmi2mid.h"
