+++ src/audio.cc
@@ -28,7 +28,6 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/log.h"
 END_EXT_C
