+++ src/freeserf.cc
@@ -38,13 +38,13 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/log.h"
   #include "src/savegame.h"
   #include "src/mission.h"
   #include "src/version.h"
   #include "src/game.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/audio.h"
 #include "src/gfx.h"
 #include "src/video-sdl.h"
