+++ src/game-init.cc
@@ -28,8 +28,8 @@
 #include "src/misc.h"
 BEGIN_EXT_C
   #include "src/mission.h"
-  #include "src/data.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/interface.h"
 #include "src/version.h"
