+++ src/gfx.cc
@@ -26,9 +26,9 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/log.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/video.h"
 
 #ifdef min
