+++ src/interface.cc
@@ -27,9 +27,9 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/debug.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/audio.h"
 #include "src/freeserf_endian.h"
 #include "src/freeserf.h"
