+++ src/popup.cc
@@ -28,10 +28,10 @@
 
 #include "src/misc.h"
 BEGIN_EXT_C
-  #include "src/data.h"
   #include "src/game.h"
   #include "src/debug.h"
 END_EXT_C
+#include "src/data.h"
 #include "src/audio.h"
 #include "src/gfx.h"
 #include "src/interface.h"
